Consolidate vessel segmentation into one validated pipeline
Folds the fast, in-process batched segmentation pipeline (previously living
separately in faster-segmentation-test/, validated at 16.15x speedup with
Dice 0.9998-1.0 vs the old pipeline) into segmentation/ as the sole
implementation. Removes the old subprocess-per-file pipeline it replaced
(run_vessel_segmentation, process_single_file, the old CLI) now that
duplication between the two locations is gone.

Also bumps segmentation/slurm/submit_batch_segmentation_array.slurm's SBATCH
resources (gpuq, 4 CPUs, 32G) to match what the batched pipeline actually
needs, since the array job still had the old serial pipeline's minimal
resource request.

validate_speed_and_accuracy.py is kept but frozen/historical: it drove the
old pipeline's --file-index CLI, which no longer exists, so it can't
actually re-run a comparison -- it documents how validation_results.md's
numbers were produced. Re-ran the CPU-safe equivalence tests
(test_batching_equiv.py, test_preprocess_parallel.py) against the
consolidated code; both still pass bit-identical.
