--- faster-segmentation-test/vanguard.code-workspace
@@ -1,7 +1,0 @@
-{
-	"folders": [
-		{
-			"path": ".."
-		}
-	]
-}
